pyproject.toml
[tool.poetry]
name = "savant"
version = "0.2.0"
description = "An intelligent optimization problem solver using LLMs and Answer Set Programming"
authors = ["Dipankar Sarkar <me@dipankar.name>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/your-username/savant"
keywords = ["optimization", "llm", "asp", "clingo", "clorm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
openai = "^1.0.0"
clorm = "^1.6.0"
click = "^8.0.0"
dspy-ai = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
savant = "savant.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
